Always use local buffer-undo-list in body of `evil-with-undo'
For some reason I do not understand the undo-boundary before the paste-command in `evil-paste-behind' is removed when called from `evil-paste-pop'. This led to the bad situation that a following `evil-paste-pop' would not only remove the latest paste but also the buffer change before.

The new implementation always works with a new empty buffer-undo-list
which ensures the existence of the undo-boundary when updating the
global buffer-undo-list.

Note that this problem only arises in Emacs 23.3 and only when
executed in a keyboard-macro (as in the tests), not when using
interactively. I assume that undo boundaries are removed during the
execution of a keyboard macro.