Add jump commands "C-o" and "C-i"
They are implemented in terms of Emacs' own mark ring, so "C-o" is
essentially the same as "C-u C-SPC". However, maybe this should be
replaced with a custom ring as the mark ring is buffer-local and
doesn't support jumping between buffers. The binding for "C-i" is
contingent on the customization variable `evil-want-C-i-jump' since
"C-i" is equivalent to "<tab>" when running Emacs in the terminal.