Reorder and rename commands
Move all operators into `evil-operator-state-map'.
Use common vi mnemonics ("a" stands for "append",
"o" for "open", etc.) for naming the commands. Rename
`evil-insert-before' to `evil-insert' ("i"),
`evil-insert-beginning-of-line' to `evil-insert-line' ("I"),
`evil-insert-after' to `evil-append' ("a"),
`evil-insert-end-of-line' to `evil-append-line' ("A"),
`evil-insert-below' to `evil-open-below' ("o"),
`evil-insert-above' to `evil-open-above' ("O"),
`evil-change-chars' to `evil-substitute' ("s"),
`evil-join-successive-lines' to `evil-join' ("J"),
`evil-move-to-window-line' to `evil-window-top' ("H"),
`evil-move-to-middle-window-line' to `evil-window-middle' ("M"),
`evil-move-to-last-window-line' to `evil-window-bottom' ("L"),
`evil-move-to-first-non-blank-end' to `evil-goto-line' ("G"), and
`evil-move-to-first-non-blank-beg' to `evil-goto-first-line' ("gg").