Add local scope for `evil-inhibit-operator'
We want the variable `evil-inhibit-operator' to work as if a
`let' form surrounds the whole body of an operator command,
including its `interactive' specification. However, it is not
allowed to put `interactive' inside `let', so the variable
`evil-inhibit-operator-value' is used to transfer the value
from one local scope to another.