Add buffer-local variables for central Visual concepts
Use separate variables instead of storing things as properties
of `evil-visual-overlay'. The overlay approach was intended to
implement encapsulation, but it amounted to little more than a
layer of indirection. The variables put the Visual logic out in
the open, where it belongs. Also rename the :injective property
to :one-to-one for clarity.