Ensure `evil-scroll-to-center' centers current line when used without count and respecting line-wrap.
When line-wrapping is enabled the current line number is not a good
measure for determining the line to be centered, if the current line
is line wrapped. This change makes `evil-scroll-to-center' not move
point before recentering if no explicit count is given. In this
situation Emacs `recenter' usually does the right thing.