Reorder window bindings for use in terminals
In terminals, "C-H" and "C-h" are not distinguised. So, by defining
the lower case versions after the upper case versions, we can default
to the less annoying option of moving in the direction expected rather
than moving the current window in that direction.

Add "C-H", "C-h" window binding to support moving left while holding
down key.