Inhibit call of `evil-ex-update' during ex completion
The completion functions called via `completion-at-point' possibly
modify the content of the minibuffer arbitrarily. The content of the
minibuffer may temporarily be an invalid ex command line and would
case `evil-ex-update' to raise an error. The new approach is to wrap
`completion-at-point' in the function
`evil-ex-run-completion-at-point' which disables `evil-ex-update'
during the completion process but calls it once the completion is
finished. Therefore `evil-ex-update' is always called with the
minibuffer content being a valid ex command line.