Add `evil-quote-range' possibly selecting white space around quote objects
This function calls `evil-paren-range' to select the object itself and
possibly adds trailing or preceding white space if appropriate. "a"
text objects select the object plus white space, "2i" (i.e., inner
with count 2) select the objects including the quotes but no
additional white space. Note that the inner quote objects use
`current-prefix-arg' the get the actual count because the `count'
parameter itself may be set to 1 or -1 by the general text-object code
(in visual state). This should be improved in future! (fixes #114)