Make :quit and friends close the current emacsclient
If a frame is opened by a new emacsclient, a :quit command (or one of
its variants :qa, :wq or :wqa) should close this client if that frame
is the one being closed. This is done by calling
`save-buffers-kill-terminal' where appropriate or examining the value
of the 'client frame parameter. This closes #143.