Ensure `evil-execute-in-emacs-state' toggles state in original buffer
If the Emacs command being executed changes the current buffer, e.g.
activates the minibuffer, then evil toggles the state of the new
buffer because this happens in a post-command-hook. This patch ensures
that the original buffer (i.e. the buffer in which
`evil-execute-in-emacs-state' has been called) is reset to its
previous evil state. This fixes #181.