Implement `evil-track-last-insert' using an `after-change-functions' hook.
The previous implementation inspected the undo-list. This may be
problematic with certain packages (e.g. undo-tree) or if no undoing is
available in the buffer. The new version uses an
after-change-functions hook instead. This is more appropriate for this
particular task (the insertion/deletion information is available at
hand) and also has a simpler implementation.