Keep mouse selected visual selection after releasing the button.
The mouse handling code deactivates the visual selection when the
button is released and point has not been moved. While this is desired
for character selection, it is not expected for word, line or block
selections. This causes, e.g., the selected line to disappear if the
button is released after triple clicking without moving point.

The reason is that the visual selection does not correspond to (point)
and (mark) in these cases. In fact, a visual line selection (and word
or block selection) should *always* be considered as if point or mark
has been changed.

This change augments the point/mark test by adding a test whether the
current visual selection type is line or block or if `click-count`
equals 1 (which corresponds to word selection).

This fixes #238.