Make `evil-read-key` support digraphs
If a character argument is read using `evil-read-key` (e.g. the
motions f F t T and the command r), a digraph can be inserted by
pressing `C-k` followed by the two digraph characters.

For this, the function `evil-read-digraph-char` is bound to `C-k` (by
default) in `evil-read-key-map`. This functions performs the reading
and transformation of the digraph as well as the overlay handling in
the buffer.

The command `evil-insert-digraph` now simply calls
`evil-read-digraph-char`.