Improve performance of `evil-resize-window` (fix #268)
In Emacs <= 23 the function `enlarge-window` may delete other windows.
Because of this `evil-resize-window` changed the size of a window in
steps of one, checking whether the window configuration changed. This
may be very inefficient.

In Emacs 24 `enlarge-window` does not delete other windows anymore, so
we can use this function directly. In Emacs 23 we use
`inhibit-redisplay` to prevent Emacs from redisplaying during the
resizing process, which should improve the performance.