fix default cursor
The default cursor value in variable `evil-default-cursor` should not
be initialised with the colour of the cursor of the current frame. On
the one hand, no current frame might be available (e.g. in Emacs
server mode), on the other hand this colour may be inappropriate if
the colour theme is changed afterwards. Instead we adapt the function
`evil-refresh-cursor` to choose the current frames colour at
runtime. This addresses merge request 52.