Flip priority of minor-mode and auxiliary maps
Since evil-define-key can be used for major-mode keys, if someone
defines all of their minor-mode keys with evil-define-minor-mode-key
they can be shadowed by the major-mode ones. Consider defining
keys for org-mode using evil-define-key and keys for org-capture-mode
using evil-define-minor-mode key. The former might shadow the latter
and there is no good way to fix this in the current priority order.