Move definition of evil-mode
evil-mode was defined last in the loading sequence, after an advice had been
defined for it, which raised a warning at load time. There does not seem to be
any reason to defer defining evil-mode to the end of the load sequence as it
just defines a minor mode and all of the constituent functions have been
defined at this point.